Move l2_with_l1_fees in evm folder
