+++ rotkehlchen/chain/base/etherscan.py
@@ -1,7 +1,7 @@
 import logging
 from typing import TYPE_CHECKING
 
-from rotkehlchen.chain.l2_with_l1_fees.etherscan import L2WithL1FeesEtherscan
+from rotkehlchen.chain.evm.l2_with_l1_fees.etherscan import L2WithL1FeesEtherscan
 from rotkehlchen.logging import RotkehlchenLogsAdapter
 from rotkehlchen.types import ExternalService, SupportedBlockchain
