+++ rotkehlchen/chain/base/transactions.py
@@ -1,7 +1,7 @@
 import logging
 from typing import TYPE_CHECKING
 
-from rotkehlchen.chain.l2_with_l1_fees.transactions import L2WithL1FeesTransactions
+from rotkehlchen.chain.evm.l2_with_l1_fees.transactions import L2WithL1FeesTransactions
 from rotkehlchen.logging import RotkehlchenLogsAdapter
 
 if TYPE_CHECKING:
