+++ rotkehlchen/chain/scroll/node_inquirer.py
@@ -5,8 +5,8 @@ from eth_typing import BlockNumber
 
 from rotkehlchen.chain.constants import DEFAULT_EVM_RPC_TIMEOUT
 from rotkehlchen.chain.evm.contracts import EvmContracts
+from rotkehlchen.chain.evm.l2_with_l1_fees.node_inquirer import L2WithL1FeesInquirer
 from rotkehlchen.chain.evm.types import string_to_evm_address
-from rotkehlchen.chain.l2_with_l1_fees.node_inquirer import L2WithL1FeesInquirer
 from rotkehlchen.constants.assets import A_ETH
 from rotkehlchen.fval import FVal
 from rotkehlchen.greenlets.manager import GreenletManager
