+++ rotkehlchen/chain/scroll/transactions.py
@@ -1,6 +1,6 @@
 from typing import TYPE_CHECKING
 
-from rotkehlchen.chain.l2_with_l1_fees.transactions import L2WithL1FeesTransactions
+from rotkehlchen.chain.evm.l2_with_l1_fees.transactions import L2WithL1FeesTransactions
 
 if TYPE_CHECKING:
     from rotkehlchen.db.dbhandler import DBHandler
